Add safe query and safe write for database
Due to high parallelism, when the database interface gets hit
with many write requests at once, it freezes up. This can be
mitigated in SQLite by checking for an OperationalError and
then retrying until the database works. This fix seems to work
well even when hundreds of write requests are hitting the database
at the same time. An improvement for the future may be to look
at the granularity of retry times.