Improve integration and merging steps
This commit takes several steps to improving the overall quality of the final builds.
It adds in FDR and Retention Time calculators which allow merged calculation of
statistics necessary to the final build. It also fixes a bug which was causing
PSMs to be reanlyzed several times in different splits.