Add initial refactor and tests for database interface
Here we are starting the refactor process for connections to the database.
Calls to the database will now be handled by a single class and standardized
methods which the pipeline needs will be handled by individual classes.
This will allow a lot cleaner deliniation of code. This commit also
contains the first of many tests that will be added to the pipeline.